+++ src/i18n/validator/index.js
@@ -1,0 +1,93 @@
+import al from './al.json';
+import ar from './ar.json';
+import bg from './bg.json';
+import ca from './ca.json';
+import cs from './cs.json';
+import da from './da.json';
+import de from './de.json';
+import el from './el.json';
+import en from './en.json';
+import es from './es.json';
+import et from './et.json';
+import eu from './eu.json';
+import fa from './fa.json';
+import fi from './fi.json';
+import fr from './fr.json';
+import he from './he.json';
+import hr from './hr.json';
+import hu from './hu.json';
+import id from './id.json';
+import it from './it.json';
+import ja from './ja.json';
+import ko from './ko.json';
+import lt from './lt.json';
+import lv from './lv.json';
+import ms from './ms.json';
+import nl from './nl.json';
+import no from './no.json';
+import pl from './pl.json';
+import pt_br from './pt_br.json';
+import pt_pt from './pt_pt.json';
+import ro from './ro.json';
+import ru from './ru.json';
+import sk from './sk.json';
+import sl from './sl.json';
+import sq from './sq.json';
+import sr from './sr.json';
+import sv from './sv.json';
+import th from './th.json';
+import tk from './tk.json';
+import tr from './tr.json';
+import ua from './ua.json';
+import uk from './uk.json';
+import ur from './ur.json';
+import zh_cn from './zh_cn.json';
+import zh_tw from './zh_tw.json';
+
+export {
+  al,
+  ar,
+  bg,
+  ca,
+  cs,
+  da,
+  de,
+  el,
+  en,
+  es,
+  et,
+  eu,
+  fa,
+  fi,
+  fr,
+  he,
+  hr,
+  hu,
+  id,
+  it,
+  ja,
+  ko,
+  lt,
+  lv,
+  ms,
+  nl,
+  no,
+  pl,
+  pt_br,
+  pt_pt,
+  ro,
+  ru,
+  sk,
+  sl,
+  sq,
+  sr,
+  sv,
+  th,
+  tk,
+  tr,
+  ua,
+  uk,
+  ur,
+  zh_cn,
+  zh_tw
+}